requirements.txt
appdirs==1.4.2
defusedxml==0.5.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10.6
django-belt==1.0a2
django-model-utils==2.6.1
django-widget-tweaks==1.4.1
future==0.16.0
gunicorn==19.6.0
oauthlib==2.0.1
packaging==16.8
psycopg2==2.7
PyJWT==1.4.2
pyparsing==2.1.10
python-social-auth==0.3.6
python-twitter==3.2.1
python3-openid==3.1.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
static3==0.7.0
